// src/content/projectiles.generated.ts
export type ProjectileId =
  | 'pistol'
  | 'assault_rifle'
  | 'shotgun'
  | 'submachine_gun'
  | 'gauss_gun'
  | 'plasma_rifle'
  | 'plasma_minigun'
  | 'pulse_gun'
  | 'ion_rifle'
  | 'ion_minigun'
  | 'ion_cannon'
  | 'shrinkifier'
  | 'blade_gun'
  | 'spider_plasma'
  | 'plasma_cannon'
  | 'splitter_gun'
  | 'plague_spreader'
  | 'rainbow_gun'
  | 'fire_bullets';

export interface ProjectileDef {
  typeId: number;
  id: ProjectileId;
  name: string;
  speed: number;
  lifeSeconds: number;
  radius: number;
  baseDamageMeta: number | null;
  damagePool: number | null;
}

export const PROJECTILES: ProjectileDef[] = [
  {
    typeId: 1,
    id: 'pistol',
    name: 'PISTOL',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 55,
    damagePool: 1,
  },
  {
    typeId: 2,
    id: 'assault_rifle',
    name: 'ASSAULT_RIFLE',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 50,
    damagePool: 1,
  },
  {
    typeId: 3,
    id: 'shotgun',
    name: 'SHOTGUN',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 60,
    damagePool: 1,
  },
  {
    typeId: 5,
    id: 'submachine_gun',
    name: 'SUBMACHINE_GUN',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 45,
    damagePool: 1,
  },
  {
    typeId: 6,
    id: 'gauss_gun',
    name: 'GAUSS_GUN',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 215,
    damagePool: 300,
  },
  {
    typeId: 9,
    id: 'plasma_rifle',
    name: 'PLASMA_RIFLE',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 30,
    damagePool: 1,
  },
  {
    typeId: 11,
    id: 'plasma_minigun',
    name: 'PLASMA_MINIGUN',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 35,
    damagePool: 1,
  },
  {
    typeId: 19,
    id: 'pulse_gun',
    name: 'PULSE_GUN',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 20,
    damagePool: 1,
  },
  {
    typeId: 21,
    id: 'ion_rifle',
    name: 'ION_RIFLE',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 15,
    damagePool: 1,
  },
  {
    typeId: 22,
    id: 'ion_minigun',
    name: 'ION_MINIGUN',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 20,
    damagePool: 1,
  },
  {
    typeId: 23,
    id: 'ion_cannon',
    name: 'ION_CANNON',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 10,
    damagePool: 1,
  },
  {
    typeId: 24,
    id: 'shrinkifier',
    name: 'SHRINKIFIER',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 45,
    damagePool: 1,
  },
  {
    typeId: 25,
    id: 'blade_gun',
    name: 'BLADE_GUN',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 20,
    damagePool: 50,
  },
  {
    typeId: 26,
    id: 'spider_plasma',
    name: 'SPIDER_PLASMA',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 10,
    damagePool: 1,
  },
  {
    typeId: 28,
    id: 'plasma_cannon',
    name: 'PLASMA_CANNON',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 10,
    damagePool: 1,
  },
  {
    typeId: 29,
    id: 'splitter_gun',
    name: 'SPLITTER_GUN',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 30,
    damagePool: 1,
  },
  {
    typeId: 41,
    id: 'plague_spreader',
    name: 'PLAGUE_SPREADER',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 15,
    damagePool: 1,
  },
  {
    typeId: 43,
    id: 'rainbow_gun',
    name: 'RAINBOW_GUN',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 10,
    damagePool: 1,
  },
  {
    typeId: 45,
    id: 'fire_bullets',
    name: 'FIRE_BULLETS',
    speed: 20,
    lifeSeconds: 1,
    radius: 1,
    baseDamageMeta: 60,
    damagePool: 240,
  },
];

export const PROJECTILE_BY_ID: Record<ProjectileId, ProjectileDef> = {
  'pistol': PROJECTILES.find(p => p.id === 'pistol')!,
  'assault_rifle': PROJECTILES.find(p => p.id === 'assault_rifle')!,
  'shotgun': PROJECTILES.find(p => p.id === 'shotgun')!,
  'submachine_gun': PROJECTILES.find(p => p.id === 'submachine_gun')!,
  'gauss_gun': PROJECTILES.find(p => p.id === 'gauss_gun')!,
  'plasma_rifle': PROJECTILES.find(p => p.id === 'plasma_rifle')!,
  'plasma_minigun': PROJECTILES.find(p => p.id === 'plasma_minigun')!,
  'pulse_gun': PROJECTILES.find(p => p.id === 'pulse_gun')!,
  'ion_rifle': PROJECTILES.find(p => p.id === 'ion_rifle')!,
  'ion_minigun': PROJECTILES.find(p => p.id === 'ion_minigun')!,
  'ion_cannon': PROJECTILES.find(p => p.id === 'ion_cannon')!,
  'shrinkifier': PROJECTILES.find(p => p.id === 'shrinkifier')!,
  'blade_gun': PROJECTILES.find(p => p.id === 'blade_gun')!,
  'spider_plasma': PROJECTILES.find(p => p.id === 'spider_plasma')!,
  'plasma_cannon': PROJECTILES.find(p => p.id === 'plasma_cannon')!,
  'splitter_gun': PROJECTILES.find(p => p.id === 'splitter_gun')!,
  'plague_spreader': PROJECTILES.find(p => p.id === 'plague_spreader')!,
  'rainbow_gun': PROJECTILES.find(p => p.id === 'rainbow_gun')!,
  'fire_bullets': PROJECTILES.find(p => p.id === 'fire_bullets')!,
};

export const PROJECTILE_BY_TYPE_ID: Record<number, ProjectileId> = {
  [1]: 'pistol',
  [2]: 'assault_rifle',
  [3]: 'shotgun',
  [5]: 'submachine_gun',
  [6]: 'gauss_gun',
  [9]: 'plasma_rifle',
  [11]: 'plasma_minigun',
  [19]: 'pulse_gun',
  [21]: 'ion_rifle',
  [22]: 'ion_minigun',
  [23]: 'ion_cannon',
  [24]: 'shrinkifier',
  [25]: 'blade_gun',
  [26]: 'spider_plasma',
  [28]: 'plasma_cannon',
  [29]: 'splitter_gun',
  [41]: 'plague_spreader',
  [43]: 'rainbow_gun',
  [45]: 'fire_bullets',
};